+++ src/directory/block.rs
@@ -154,7 +154,7 @@ where
     fn next(&mut self) -> Option<Self::Item> {
         while !self.stopped && self.pos + 8 <= BLK_SIZE {
             if let Ok(entry) = self.inner.read_entry_at(self.pos) {
-                self.pos += entry.get_entry_len() as usize;
+                self.pos += entry.get_entry_len();
                 if entry.get_inode() != crate::types::InodeNumber(0) {
                     return Some(Ok(entry));
                 }
@@ -248,7 +248,7 @@ where
             {
                 let (ppos, pos) = (self.ppos, self.pos);
                 self.ppos = Some(self.pos);
-                self.pos += entry.get_entry_len() as usize;
+                self.pos += entry.get_entry_len();
                 Some(Ok(DirectoryBlockIteratorMutCursor {
                     block: self.block,
                     ppos,
